requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
Flask-CORS==4.0.0
SQLAlchemy==1.4.54
python-dotenv==1.0.0
twilio==8.10.0
gunicorn==21.2.0
Werkzeug==2.3.7

